fix(wiki): prune score float64 widening + isVerbose strict threshold
- wiki_article.go: widen WordCount and DaysUnread to float64 individually
  before multiplying — consistent with the denominator formula and avoids
  potential int overflow on very large values
- WikiCatalog.tsx: change isVerbose to use > instead of >= so the
  boundary entry at exactly the 90th-percentile cutoff does not earn
  the badge; only entries strictly above it do

+++ internal/team/wiki_article.go
@@ -233,7 +233,7 @@ func (r *Repo) BuildCatalog(ctx context.Context, sortBy string, readLog *ReadLog
 		e := &entries[i]
 		if e.WordCount > 0 {
 			denom := math.Max(float64(e.HumanReadCount)+0.3*float64(e.AgentReadCount), 1.0)
-			e.PruneScore = float64(e.WordCount*e.DaysUnread) / denom
+			e.PruneScore = float64(e.WordCount) * float64(e.DaysUnread) / denom
 		}
 	}
